notification_bridge/requirements.txt
aiokafka
httpx
python-dotenv
PyYAML